\name{adjacency}
\alias{adjacency}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
An example of an adjacency matrix.
}
\description{
An example of an adjacency matrix with 1000 genes randomly generated from TCGA KIRP gene expression data.
}


\references{
DiSNEP: a Disease-Specific gene Network Enhancement to improve Prioritizing candidate disease genes (2020), Peifeng Ruan, Shuang Wang.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\examples{

#An example of an adjacency matrix
data(adjacency)

}
